Compiler warning cleanup.

+++ driver_ais.c
@@ -682,7 +682,7 @@ bool ais_binary_decode(const int debug,
 		UCHARS(56, ais->type8.dac200fid24.country);
 #define ARRAY_BASE 68
 #define ELEMENT_SIZE 25
-		for (i = 0; ARRAY_BASE + (ELEMENT_SIZE*i) < bitlen; i++) {
+		for (i = 0; ARRAY_BASE + (ELEMENT_SIZE*i) < (int)bitlen; i++) {
 		    int a = ARRAY_BASE + (ELEMENT_SIZE*i);
 		    ais->type8.dac200fid24.gauges[i].id = UBITS(a+0,  11);
 		    ais->type8.dac200fid24.gauges[i].level = SBITS(a+11, 14);
